requirements.txt
telethon>=1.36.0
pandas>=2.2.0
numpy>=1.26.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
llama-cpp-python>=0.2.0
